Elegant Solution
Added an elegant solution to solving a Sudoku Board.  Only checks for
squares with single elements, or elements that are found only once in a
row, column, or box.  Using the Sudoku Board 1-6, performance
improvement for easy to moderate problems is ~5-6 times faster, if not
higher.  Performance improvement for difficult problems is negligible.