Refactoring
Updated Engine to be more readable.  Used variables in hidden_double to
make code more readable in the future.  Updated hidden double to return
true if any possible values were removed instead of returning true if a
hidden double was found.  Updated Cell class to return if a possible
value was removed.