Refactoring
Updated the double calls to use a common function to find doubles in
cells.  Instead of the double_row, double_column, and double_box being
unique, each makes a call to two a find_double and process_double.  This
will make future updates easier to apply as it will be in one location
instead of three.  Also added comments to make the code clearer